Update environment.go to introduce GLLVM_BUILD_PRODUCTS
This introduces GLLVM_BUILD_PRODUCTS as a way of specifying a file that will contain all the build products and their dependencies. Right now, gllvm is able to record all compilation and linking steps. But for any given package, the build is not just an executable or library. It can also be tests that are not part of the build product of a package. This allows us to capture all builds, including tests.